fixed issue with MySQL

+++ src/main/java/at/pavlov/cannons/dao/DeleteCannonTask.java
@@ -31,13 +31,13 @@ public class DeleteCannonTask extends BukkitRunnable{
     public void run() {
         try (Statement statement = Cannons.getPlugin().getConnection().createStatement()) {
             if (cannonId == null && playerId == null){
-                statement.executeUpdate(String.format("DELETE FROM '%s'", Cannons.getPlugin().getCannonDatabase()));
+                statement.executeUpdate(String.format("DELETE FROM %s", Cannons.getPlugin().getCannonDatabase()));
             }
             else if (cannonId != null) {
-                statement.executeUpdate(String.format("DELETE FROM '%s' WHERE id='%s'", Cannons.getPlugin().getCannonDatabase(), cannonId));
+                statement.executeUpdate(String.format("DELETE FROM %s WHERE id=%s", Cannons.getPlugin().getCannonDatabase(), cannonId));
             }
             else{
-                statement.executeUpdate(String.format("DELETE FROM '%s' WHERE owner='%s'", Cannons.getPlugin().getCannonDatabase(), playerId));
+                statement.executeUpdate(String.format("DELETE FROM %s WHERE owner=%s", Cannons.getPlugin().getCannonDatabase(), playerId));
             }
         } catch (Exception e) {
             e.printStackTrace();
